Put new reports view behind projects feature flag
